requirements.txt
pyrebase4
streamlit
langchain
faiss-cpu
sentence-transformers
transformers


